Move setting of states and counties into monthly emissions reducer, and update method of setting it to use emissionsChanges data returned in results reducer
